tests/wpt/web-platform-tests/html/semantics/forms/the-selectmenu-element/selectmenu-form-elements.tentative.html

<!DOCTYPE html>
<html lang="en">
<title>HTMLSelectMenuElement Test: form.elements</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<form id="form0">
<selectmenu id="selectmenu0">
  <option>one</option>
  <option>two</option>
  <option>three</option>
</selectmenu>
<form>

<script>
promise_test(async t => {
  // TODO: Move test to /the-form-element/form-elements-filter.html once
  // <selectmenu> becomes part of the HTML spec.
  const formElements = document.querySelector("#form0").elements;
  assert_equals(formElements.length, 1);
  assert_equals(formElements[0].id, "selectmenu0");
}, "Check that <selectmenu> is exposed in form.elements");
</script>
